Removed ReportHandlerMixIn methods from plugin registry
